# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Charity market API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:80"
        }
    ],
    "paths": {
        "/auth/registration": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Регистраци пользователя и получение токена",
                "operationId": "authRegister",
                "requestBody": {
                    "description": "Данные пользователя",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "login",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "login": {
                                        "type": "string",
                                        "example": "user1"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "qwer1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Регистрация прошла успешно",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Аутентификация пользователя и получение токена",
                "operationId": "authLogin",
                "requestBody": {
                    "description": "Данные пользователя",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "qwer1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Аутентификация прошла успешно",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Выход из системы (деактивация всех токенов пользователя)",
                "operationId": "authLogout",
                "responses": {
                    "200": {
                        "description": "Все хорошо"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/items/all": {
            "get": {
                "tags": [
                    "Items"
                ],
                "summary": "Получение всех существующих товаров с пагинацией",
                "operationId": "itemsAll",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер страницы",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Все хорошо"
                    },
                    "500": {
                        "description": "Ошибка на стороне сервера"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/items/add": {
            "post": {
                "tags": [
                    "Items"
                ],
                "summary": "Добавление товара авторизированным пользователем",
                "operationId": "itemsAdd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Наименование",
                                        "type": "string"
                                    },
                                    "cat_id": {
                                        "description": "Id категории",
                                        "type": "integer"
                                    },
                                    "desc": {
                                        "description": "Описание",
                                        "type": "string"
                                    },
                                    "price": {
                                        "description": "Цена",
                                        "type": "string"
                                    },
                                    "photo": {
                                        "description": "Фотография",
                                        "type": "file",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Товар добавлен",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}